+++ src/api/getKeysOfObjectsToDelete.ts
@@ -1,6 +1,12 @@
 import {ObjectDiffStatusMap, ObjectDiffStatus} from './types';
 
-export function getKeysOfObjectsToDelete(diff: ObjectDiffStatusMap): string[] {
+export function getKeysOfObjectsToDelete(
+  diff: ObjectDiffStatusMap,
+  shouldDeleteDeletedObjects: boolean
+): string[] {
+  if (!shouldDeleteDeletedObjects) {
+    return [];
+  }
   return Object.keys(diff).reduce((accum, key) => {
     if (diff[key] === ObjectDiffStatus.DELETED) {
       accum = [...accum, key];